refactor(agent): 抽出 owned presenter trace 终结器，base.py 回到 1000 行限内
ecf14458 加入的 complete_owned_presenter_trace 把 base.py 推到 1025 行，
超出 largeBackendModulesSource 的 1000 行上限（CI 会挂）。原样移到
src/agents/core/trace_finalizer.py，base.py 以导入方式继续暴露同名函数
（测试的 from src.agents.core.base import 路径不变），行为零变化。
